@ngimel 2021-05-26
@peterbell10 android tests are failing with
```
Test Case ModelRunner
* Running ModelRunner.basic
***** Failure in unknown file
C++ exception with description "dtype 'Bool' not selected for kernel tag "nonzero_count_cpu"
  
```