@peterbell10 2021-05-26
Is there a list of "selected" kernels somewhere? I'm adding new kernels, so I assume those need to be added to whatever list.