@ngimel 2021-05-27
Sorry, that suggestion still didn't work. I'm working with mobile team to try and get it resolved.