@ngimel 2021-05-28
Sorry @peterbell10, xla error is real (probably related to new dispatch for count_nonzeros), reverting